Release v1.0.0
Cells Client is an easy to use command line tool to communicate with a Pydio Cells v2 server. It uses OAuth2 for authentication, avoiding any credentials to be stored on the machine, and provides basic operations for listing, downloading and uploading files to the server. More to come very soon.